responsive styling complete and all errors complete
