+++ pages/create-note.php
@@ -9,27 +9,22 @@
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
-        require_once '../include/database.php';
-        require_once '../include/check-authorisation.php';
-        checkMachineIdIsSet($conn);
-        require_once '../include/menu.php';
-        setBannerColour($conn);
+        require_once '../include/page-defaults.php';
     ?>
     <div id=body-container>
     <?php 
     
     $jobID = $_GET['jobID'];
     
-    echo "<div class='headerContainer'>";
+    echo "<div class='header-container'>";
     echo "<h1>Create Task Note</h1>";
-    echo "<div class='headerLinks'>";
+    echo "<div class='header-links'>";
     echo "<a href='job.php?id=" . $jobID . "&machineID=" . $_GET['machineID']. "'>Cancel</a>";
     echo "</div>";
     echo "</div>";
-    echo "<div class='createJobForm'>";
+    echo "<div class='create-job-form'>";
     echo "<form action='../system/create-note.php?jobID=" . $_GET['jobID'] . "&machineID=" . $_GET['machineID'] . "' method='POST'>";
-    echo "<div class='innerJobForm'>";
+    echo "<div class='inner-job-form'>";
     echo "<label for='category'>Category:</label> ";
     echo "<div class='select-dropdown'>";
     echo "<select id='category' name='category' required>";
@@ -43,7 +38,7 @@
     echo "</select>";
     echo "</div>";
     echo "<label for='priority'>Priority:</label>";
-    echo "<div class='select-dropdown'>";
+    echo "<div class='select-dropdown note-priority'>";
     echo "<select id='priority' name='priority' required>";
     echo "<option value='1'>Low</option>";
     echo "<option value='2'>Medium</option>";
@@ -51,7 +46,7 @@
     echo "</select>";
     echo "</div>";
     echo "</div>";
-    echo "<div class='innerJobForm'>";
+    echo "<div class='inner-job-form'>";
     echo "<label for='description'>Description:</label>";
     echo "<textarea id='description' name='description' required></textarea>";
     echo "<input name='submit' type='submit' value='Submit' />";
