+++ pages/note.php
@@ -6,16 +6,11 @@
     <meta name="author" content="Group 18" />
     <link rel="stylesheet" href="../styles/style.css">
     <script src="../scripts/banner.js"></script>
-    <script src="../scripts/deleteJob.js"></script>
 </head>
 <body>
     <?php
-        require_once '../include/functions.php';
-        require_once '../include/database.php';
-        require_once '../include/check-authorisation.php';
-        checkMachineIdIsSet($conn);
-        require_once '../include/menu.php';
-        setBannerColour($conn);
+        require_once '../include/page-defaults.php';
+        require_once '../scripts/jobs.php';
     ?>
     <div id=body-container>
     <?php
