+++ sql/insert-jobs.sql
@@ -1,0 +1,80 @@
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Please print out 300 figures",
+    "1",
+    "2",
+    "2",
+    "2024-09-25 09:15:32",
+    "In progress",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "This machine needs to be fixed, it seems to have stopped operating",
+    "3",
+    "2",
+    "3",
+    "2024-09-25 10:04:13",
+    "Delayed",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Ensure that this machine is operational for next week",
+    "6",
+    "2",
+    "1",
+    "2024-09-25 14:53:27",
+    "Awaiting confirmation",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Please be operating this today",
+    "4",
+    "3",
+    "2",
+    "2024-09-25 08:33:18",
+    "Completed",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "I need you to cover this shift as soon as possible",
+    "7",
+    "3",
+    "3",
+    "2024-09-25 13:48:09",
+    "In progress",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Going to start you off on something easy for your first day back from injury, work with Rob on the 3D printer",
+    "1",
+    "4",
+    "1",
+    "2024-09-25 08:48:10",
+    "In progress",
+    "0"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Fix this machine please",
+    "9",
+    "2",
+    "3",
+    "2024-09-24 08:30:28",
+    "Completed",
+    "1"
+);
+
+INSERT INTO Job (description, machineID, OperatorID, priority, timeUpdated, status, completed) VALUES (
+    "Help Tim please",
+    "2",
+    "2",
+    "1",
+    "2024-09-23 010:11:48",
+    "Completed",
+    "1"
+);
+
