+++ sql/insert-notes.sql
@@ -1,0 +1,71 @@
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "1",
+    "Quality concern",
+    "3",
+    "2024-09-25 09:25:12",
+    "Printing out 300 figures today is going to be difficult, doing so will reduce the quality of product",
+    "2"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "1",
+    "Quality concern",
+    "3",
+    "2024-09-25 09:33:31",
+    "I don't care, get it done",
+    "1"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "2",
+    "Missing part",
+    "3",
+    "2024-09-25 11:03:52",
+    "Narrowed the issue down, but we are missing the part required to fix it, I have placed an order for the part",
+    "2"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "4",
+    "Machine issue",
+    "3",
+    "2024-09-25 11:56:31",
+    "There is a loud ticking sound when operating this machine",
+    "3"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "4",
+    "Machine issue",
+    "1",
+    "2024-09-25 12:14:01",
+    "I sorted this issue out, dont worry",
+    "3"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "5",
+    "Safety concern",
+    "3",
+    "2024-09-25 14:01:01",
+    "There are sparks flying out of this machine!",
+    "3"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "5",
+    "Safety concern",
+    "1",
+    "2024-09-25 14:05:41",
+    "Drink a cup of cement and toughen up!",
+    "1"
+);
+
+INSERT INTO Note (jobID, category, priority, timeCreated, description, personID) VALUES (
+    "6",
+    "Issue with job details",
+    "1",
+    "2024-09-25 09:04:51",
+    "This isn't really a two person job but ok",
+    "4"
+);
